added header and nav bar
added a header than includes 3 items, from left to right: an empty space item that is a <p> tag, a title which is a <p> tag, and a <div> that includes a <p> tag and an <img> tag that is going to act as the shopping cart link to checkout chosen items. The nav bar current just includes 3 empty buttons. These changes have been applied to all HTML pages